+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add support for DSYNC RR type.
+- Add support for EID, NIMLOC and DSYNC RR types.
 
 ### Fixed